pyproject.toml
[project]
name = "jp-translation"
version = "0.1.0"
description = "Markdown file translator using GCP Gemini API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coverage>=7.13.4",
    "google-genai>=1.64.0",
    "hypothesis>=6.151.9",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
